requirements.txt
langgraph
langchain-core
langchain-community
langchain-openai
python-dotenv
jupyter
ipykernel
notebook
tavily-python
typing
dataclasses
tavily